destroy_dump(): don't log "Removing" if no dump
Check for dump file existence and return success if
there is no file. This way we'll avoid printing

	Removing CT dump /vz/dump/Dump.121

in case there is no such file.

+++ src/lib/destroy.c
@@ -238,6 +238,8 @@ int destroy_dump(envid_t veid, const char *dumpdir)
 	char buf[128];
 
 	get_dump_file(veid, dumpdir, buf, sizeof(buf));
+	if (stat_file(buf) == 0)
+		return 0;
 
 	logger(1, 0, "Removing CT dump %s", buf);
 	if (unlink(buf) == 0)
